# swagger_server/swagger/swagger.yaml
openapi: 3.0.2
info:
  title: Turing Pizza API
  description: The APIs for Turing Pizza project, which includes the access to pizza
    sizes, pizza toppings, pizza specials, pizzas, stores and orders information through
    various of endpoints and methods. <br></br>You can find out more about Turing
    Pizza API at [GitHub](https://github.ccs.neu.edu/cs5500f19/TuringPizza)
  termsOfService: http://swagger.io/terms/
  contact:
    name: Turing Pizza API
    url: https://github.ccs.neu.edu/cs5500f19/TuringPizza
    email: turingpizza@turingpizza.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
- url: https://virtserver.swaggerhub.com/TuringPizza/TuringPizza/1.0.0
  description: SwaggerHub API Auto Mocking
- url: http://localhost:8080/TuringPizza/TuringPizza/1.0.0
  description: Development local server
- url: https://turingpizza-wenbo.herokuapp.com/TuringPizza/TuringPizza/1.0.0
  description: Production server
- url: https://agile-caverns-97220.herokuapp.com/TuringPizza/TuringPizza/1.0.0
  description: Production server - Maitri
- url: https://murmuring-depths-61941.herokuapp.com/TuringPizza/TuringPizza/1.0.0
  description: Production server - Steniz  
...
paths:
...
  /orders:
    get:
      tags:
      - Pizza Orders
      summary: Get all pizza orders stored on the database.
      description: 'Access to all pizza orders in the database. This will return an
        array of pizza orders objects, including each pizza order''s unique id, total_price,
        an array of pizza_id, and pizza_store_id associated with this order. '
      operationId: get_pizza_orders
      responses:
        200:
          description: 'Return all the available orders objects in an array. '
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
                x-content-type: application/json
        400:
          description: 'Bad request. Please try again later. '
      x-openapi-router-controller: swagger_server.controllers.pizza_orders_controller
    post:
      tags:
      - Pizza Orders
      summary: Create a new pizza order.
      description: 'Make a new pizza order object with unique pizza order id, total_price,
        an empty array of pizzas_id and pizza_store_id, and post it to the server. Return
        newly added pizza order object if succeeds posting, otherwise return with
        error code. When creating a new order, the pizza id list is by default empty list,
        since pizzas will be created after the initialization of the order and then all 
        pizza ids will be filled in the id list of this order. Please ignore the pizza id
        list in the example response. Also the total price will be 0 by default for the 
        same reason. '
      operationId: add_pizza_order
      requestBody:
        $ref: '#/components/requestBodies/PizzaOrderBody'
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
      x-openapi-router-controller: swagger_server.controllers.pizza_orders_controller
  /orders/{pizzaOrderId}:
    get:
      tags:
      - Pizza Orders
      summary: 'Get a pizza order given its Id. '
      description: 'Access to a pizza order with unique id, total_price, an array
        of pizzas_id, and pizza_store_id. Return the queried pizza order object. '
      operationId: get_pizza_order
      parameters:
      - name: pizzaOrderId
        in: path
        description: 'The Id of the order to be queried. '
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
          default: 123
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: 'Invalid Id supplied. '
        404:
          description: 'No pizza order found. '
      x-openapi-router-controller: swagger_server.controllers.pizza_orders_controller
    delete:
      tags:
      - Pizza Orders
      summary: Delete an order.
      description: 'Delete an order object with unique pizza order id, order price
        and post it to the server. Return deleted pizza order object if succeeds,
        otherwise return with error code. All pizzas associated to this order will 
        also be deleted since all pizzas created for this order are unique. '
      operationId: delete_order
      parameters:
      - name: pizzaOrderId
        in: path
        description: 'The Id of the order to be queried. '
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
          default: 123
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
      x-openapi-router-controller: swagger_server.controllers.pizza_orders_controller
  /orders/{pizzaOrderId}/pizzas:
    get:
      tags:
      - Pizza Orders
      summary: Get all customized pizzas for the order given the order Id.
      description: 'Access to all customized pizzas a customer has placed in the order
        by the order Id. This will return a JSON object, which is an array of Pizzas
        object. Successful return will give 200 status code, otherwise will give error
        code with error message. '
      operationId: get_order_pizzas
      parameters:
      - name: pizzaOrderId
        in: path
        description: 'The Id of the order to be queried. '
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
          default: 123
      responses:
        200:
          description: 'Return all the available customized pizzas objects in an array. '
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pizza'
                x-content-type: application/json
        400:
          description: 'Bad request. Please try again later. Not a valid order Id. '
        404:
          description: 'No pizza order found. '
      x-openapi-router-controller: swagger_server.controllers.pizza_orders_controller
    post:
      tags:
      - Pizza Orders
      summary: 'Create a new customized pizza for a specific order given the order
        Id. '
      description: 'Make a new customized pizza object with unique pizza id, pizza_special_id,
        an array of pizza_toppings_id, count, total_price and pizza_size_id to the
        server. Return newly added customized pizza object if succeeds posting, otherwise
        return with error code. '
      operationId: add_order_pizza
      parameters:
      - name: pizzaOrderId
        in: path
        description: 'The Id of the order to be queried. '
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
          default: 123
      requestBody:
        $ref: '#/components/requestBodies/PizzaBody'
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pizza'
      x-openapi-router-controller: swagger_server.controllers.pizza_orders_controller
  /orders/{pizzaOrderId}/pizza/{pizzaId}:
    get:
      tags:
      - Pizza Orders
      summary: Get access to a specific pizza in a specific Order.
      description: 'Get the information from a pizza that is available in an Order. 
        Return the queried pizza object. '
      operationId: get_pizza_in_orders
      parameters:
      - name: pizzaOrderId
        in: path
        description: 'The Id of the order to be queried. '
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
          default: 123
      - name: pizzaId
        in: path
        description: 'The Id of the customized pizza to be queried. '
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
          default: 123
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pizza'
        400:
          description: 'Bad request. '
        404:
          description: 'No pizza order found or no pizza found for the order. '
      x-openapi-router-controller: swagger_server.controllers.pizza_orders_controller
    put:
      tags:
      - Pizza Orders
      summary: Update a specific pizza in a specific Order.
      description: 'Update a pizza that is available in an Order.  All pizza orders
        will include the unique order id as an integer, the order price as float,
        unique pizza store id and unique pizza ids. Order will offer other CRUD operations
        as well. Return the updated pizza object. '
      operationId: update_pizza_in_orders
      parameters:
      - name: pizzaOrderId
        in: path
        description: 'The Id of the order to be queried. '
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
          default: 123
      - name: pizzaId
        in: path
        description: 'The Id of the customized pizza to be queried. '
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
          default: 123
      requestBody:
        $ref: '#/components/requestBodies/PizzaBody'
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pizza'
        400:
          description: 'Bad request. '
        404:
          description: 'No pizza order found or no pizza found for the order. '
      x-openapi-router-controller: swagger_server.controllers.pizza_orders_controller
    delete:
      tags:
      - Pizza Orders
      summary: Delete a pizza in order.
      description: 'Delete a pizza in order object from the server. Return deleted pizza object if succeeds,
        otherwise return with error code. '
      operationId: delete_pizza_in_orders
      parameters:
      - name: pizzaOrderId
        in: path
        description: 'The Id of the order to be queried. '
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
          default: 123
      - name: pizzaId
        in: path
        description: 'The Id of the customized pizza to be queried. '
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
          default: 123
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pizza'
        400:
          description: 'Bad request. '
        404:
          description: 'No pizza order found or no pizza found for the order. '
      x-openapi-router-controller: swagger_server.controllers.pizza_orders_controller
  # /orders/getdeliverytime:
  /orders/getcount:
    get:
      tags:
      - Pizza Orders
      summary: 'Get the proper count of pizzas using a predefined algorithm. '
      description: 'User will input the number of adults and number of children for
        the order. Both numbers are required. This will return the suggested number
        of pizzas to order broken down to amount of different sizes based on a predefined algorithm.
        1 Large Pizza = 1.25 Medium Pizza = 1.75 Small Pizza'
      operationId: get_pizza_count
      parameters:
      - name: adult
        in: query
        description: 'The number of adults for this order. '
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int64
          example: 5
      - name: child
        in: query
        description: 'The number of children for this order. '
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int64
          example: 5
      responses:
        200:
          description: 'Return the suggested number of pizzas broken down to different sizes. '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inline_response_200'
        400:
          description: 'Invalid input, please enter number of adults and children. '
      x-openapi-router-controller: swagger_server.controllers.pizza_orders_controller
    get:
      tags:
      - Pizza Orders
      summary: 'Get the proper count of pizzas using a predefined algorithm. '
      description: 'User will input the number of adults and number of children for
        the order. Both numbers are required. This will return the suggested number
        of pizzas to order basing on a predefined algorithm. '
      operationId: get_pizza_count
      parameters:
      - name: adult
        in: query
        description: 'The number of adults for this order. '
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int64
          example: 5
      - name: child
        in: query
        description: 'The number of children for this order. '
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int64
          example: 5
      responses:
        200:
          description: 'Return the suggested number of medium pizzas. '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        400:
          description: 'Invalid input, please enter number of adults and children. '
      x-openapi-router-controller: swagger_server.controllers.pizza_orders_controller
...
components:
  schemas:
...
    Pizza:
      title: Pizza
      type: object
      properties:
        pizzas_id:
          type: integer
          description: 'The id of a customized pizza in an order. '
          format: int64
          example: 123
        pizza_special_id:
          type: integer
          description: 'The id of the pizza special that this customized pizza created
            upon. '
          format: int64
          example: 123
        pizza_toppings_id:
          type: array
          description: 'The IDs of pizza toppings that this customized pizza created
            with. '
          example:
          - 1
          - 2
          - 3
          items:
            type: integer
            format: int64
        pizza_size_id:
          type: integer
          description: 'The id of the pizza size of the customized pizza. '
          format: int64
          example: 123
        count:
          type: integer
          description: 'The count of this customized pizza in an order. '
          format: int64
          example: 5
        total_price:
          type: number
          description: 'The total price of this customized pizza in an order, which
            is the single price of this customized pizza * count. '
          format: float
          example: 29.99
      example:
        total_price: 29.99
        pizza_size_id: 123
        count: 5
        pizza_toppings_id:
        - 1
        - 2
        - 3
        pizza_special_id: 123
        pizzas_id: 123
    Order:
      title: Order
      type: object
      properties:
        orders_id:
          type: integer
          description: 'The id of an order. '
          format: int64
          example: 123
        pizza_store_id:
          type: integer
          description: 'The id of a pizza store this order belongs to. '
          format: int64
          example: 123
        pizzas_id:
          type: array
          description: 'The IDs of customized pizzas that this order includes. '
          example:
          - 1
          - 2
          - 3
          items:
            type: integer
            format: int64
        total_price:
          type: number
          description: 'The total price for this order, which is the sum of the price
            of the customized pizzs within this order. This price is a pretax price. '
          format: float
          example: 29.99
      example:
        orders_id: 123
        total_price: 29.99
        pizza_store_id: 123
        pizzas_id:
        - 1
        - 2
        - 3
    inline_response_200:
      type: object
      properties:
        count:
          type: integer
          description: 'The suggested number of medium sized pizzas. '
          format: int64
          example: 5
      example:
        count: 5
...
  requestBodies:
...
    PizzaBody:
      description: Pizza object that needs to be stored in database
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Pizza'
      required: true
...
    PizzaOrderBody:
      description: Order object that needs to be stored in database
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Order'
      required: true
...